requirements.txt
asgiref==3.6.0
async-generator==1.10
attrs==23.1.0
Brotli==1.0.9
certifi==2022.12.7
cffi==1.15.1
colorama==0.4.6
crispy-bootstrap4==2022.1
cssselect2==0.7.0
Django==4.2
django-crispy-forms==2.0
exceptiongroup==1.1.1
fonttools==4.39.3
h11==0.14.0
html5lib==1.1
idna==3.4
outcome==1.2.0
Pillow==9.5.0
pycparser==2.21
pydyf==0.6.0
pyphen==0.14.0
pypng==0.20220715.0
PySocks==1.7.1
qrcode==7.4.2
reportlab==3.6.12
selenium==4.9.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
sqlparse==0.4.4
tinycss2==1.2.1
trio==0.22.0
trio-websocket==0.10.2
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.15
weasyprint==58.1
webencodings==0.5.1
wsproto==1.2.0
zopfli==0.2.2
